+++ src/Adapter/MSTest.TestAdapter/Resources/Resource.resx
@@ -267,8 +267,8 @@ Error: {1}</value>
     <value>The ExpectedException attribute defined on test method {0}.{1} threw an exception during construction.
 {2}</value>
   </data>
-  <data name="UTA_MultipleExpectedExceptionsOnTestMethod" xml:space="preserve">
-    <value>The test method {0}.{1} has multiple attributes derived from ExpectedExceptionBaseAttribute defined on it. Only one such attribute is allowed.</value>
+  <data name="UTA_MultipleAttributesOnTestMethod" xml:space="preserve">
+    <value>The test method {0}.{1} has multiple attributes derived from '{2}' defined on it. Only one such attribute is allowed.</value>
   </data>
   <data name="LegacyScenariosNotSupportedWarning" xml:space="preserve">
     <value>Warning : A testsettings file or a vsmdi file is not supported with the MSTest V2 Adapter.</value>
