+++ src/Adapter/MSTest.TestAdapter/Resources/xlf/Resource.es.xlf
@@ -171,6 +171,11 @@ pero recibió {4} argumento(s), con los tipos "{5}".</target>
         <target state="translated">No se puede crear una instancia de la clase {0}. Error: {1}.</target>
         <note></note>
       </trans-unit>
+      <trans-unit id="UTA_MultipleAttributesOnTestMethod">
+        <source>The test method {0}.{1} has multiple attributes derived from '{2}' defined on it. Only one such attribute is allowed.</source>
+        <target state="new">The test method {0}.{1} has multiple attributes derived from '{2}' defined on it. Only one such attribute is allowed.</target>
+        <note />
+      </trans-unit>
       <trans-unit id="UTA_NoValidConstructor">
         <source>Cannot find a valid constructor for test class '{0}'. Valid constructors are 'public' and either parameterless or with one parameter of type 'TestContext'.</source>
         <target state="translated">No se encuentra un constructor válido para la clase de prueba '{0}'. Los constructores válidos son 'public' y sin parámetros o con un parámetro de tipo 'TestContext'.</target>
@@ -360,11 +365,6 @@ Error: {1}</target>
 {2}</source>
         <target state="translated">El atributo ExpectedException definido en el método de prueba {0}.{1} inició una excepción durante la construcción.
 {2}</target>
-        <note></note>
-      </trans-unit>
-      <trans-unit id="UTA_MultipleExpectedExceptionsOnTestMethod">
-        <source>The test method {0}.{1} has multiple attributes derived from ExpectedExceptionBaseAttribute defined on it. Only one such attribute is allowed.</source>
-        <target state="translated">El método de prueba {0}.{1} tiene definidos varios atributos derivados de ExpectedExceptionBaseAttribute. Solo se permite uno de estos atributos.</target>
         <note></note>
       </trans-unit>
       <trans-unit id="LegacyScenariosNotSupportedWarning">
