+++ src/Adapter/MSTest.TestAdapter/Resources/xlf/Resource.fr.xlf
@@ -171,6 +171,11 @@ mais a reçu {4} argument(s), avec les types « {5} ».</target>
         <target state="translated">Impossible de créer une instance de la classe {0}. Erreur : {1}.</target>
         <note></note>
       </trans-unit>
+      <trans-unit id="UTA_MultipleAttributesOnTestMethod">
+        <source>The test method {0}.{1} has multiple attributes derived from '{2}' defined on it. Only one such attribute is allowed.</source>
+        <target state="new">The test method {0}.{1} has multiple attributes derived from '{2}' defined on it. Only one such attribute is allowed.</target>
+        <note />
+      </trans-unit>
       <trans-unit id="UTA_NoValidConstructor">
         <source>Cannot find a valid constructor for test class '{0}'. Valid constructors are 'public' and either parameterless or with one parameter of type 'TestContext'.</source>
         <target state="translated">Impossible de trouver un constructeur valide pour la classe de test « {0} ». Les constructeurs valides sont « publics » et sans paramètre ou avec un paramètre de type « TestContext ».</target>
@@ -360,11 +365,6 @@ Erreur : {1}</target>
 {2}</source>
         <target state="translated">L'attribut ExpectedException défini dans la méthode de test {0}.{1} a levé une exception durant la construction.
 {2}</target>
-        <note></note>
-      </trans-unit>
-      <trans-unit id="UTA_MultipleExpectedExceptionsOnTestMethod">
-        <source>The test method {0}.{1} has multiple attributes derived from ExpectedExceptionBaseAttribute defined on it. Only one such attribute is allowed.</source>
-        <target state="translated">Plusieurs attributs dérivés de ExpectedExceptionBaseAttribute sont définis dans la méthode de test {0}.{1}. Un seul attribut de ce type est autorisé.</target>
         <note></note>
       </trans-unit>
       <trans-unit id="LegacyScenariosNotSupportedWarning">
