+++ src/Adapter/MSTest.TestAdapter/Resources/xlf/Resource.ja.xlf
@@ -172,6 +172,11 @@ but received {4} argument(s), with types '{5}'.</source>
         <target state="translated">クラス {0} のインスタンスを作成できません。エラー: {1}。</target>
         <note></note>
       </trans-unit>
+      <trans-unit id="UTA_MultipleAttributesOnTestMethod">
+        <source>The test method {0}.{1} has multiple attributes derived from '{2}' defined on it. Only one such attribute is allowed.</source>
+        <target state="new">The test method {0}.{1} has multiple attributes derived from '{2}' defined on it. Only one such attribute is allowed.</target>
+        <note />
+      </trans-unit>
       <trans-unit id="UTA_NoValidConstructor">
         <source>Cannot find a valid constructor for test class '{0}'. Valid constructors are 'public' and either parameterless or with one parameter of type 'TestContext'.</source>
         <target state="translated">テスト クラス '{0}' の有効なコンストラクターが見つかりません。有効なコンストラクターは、'public' で、パラメーターがないもの、または 'TestContext' 型のパラメーター 1 個を取るものです。</target>
@@ -361,11 +366,6 @@ Error: {1}</source>
 {2}</source>
         <target state="translated">テスト メソッド {0}.{1} に定義されている ExpectedException 属性が、作成中に例外をスローしました。
 {2}</target>
-        <note></note>
-      </trans-unit>
-      <trans-unit id="UTA_MultipleExpectedExceptionsOnTestMethod">
-        <source>The test method {0}.{1} has multiple attributes derived from ExpectedExceptionBaseAttribute defined on it. Only one such attribute is allowed.</source>
-        <target state="translated">テスト メソッド {0}.{1} には、ExpectedExceptionBaseAttribute から派生した属性が複数定義されています。このような属性は 1 つしか許可されません。</target>
         <note></note>
       </trans-unit>
       <trans-unit id="LegacyScenariosNotSupportedWarning">
