+++ src/Adapter/MSTest.TestAdapter/Resources/xlf/Resource.ko.xlf
@@ -171,6 +171,11 @@ but received {4} argument(s), with types '{5}'.</source>
         <target state="translated">{0} 클래스의 인스턴스를 만들 수 없습니다. 오류: {1}</target>
         <note></note>
       </trans-unit>
+      <trans-unit id="UTA_MultipleAttributesOnTestMethod">
+        <source>The test method {0}.{1} has multiple attributes derived from '{2}' defined on it. Only one such attribute is allowed.</source>
+        <target state="new">The test method {0}.{1} has multiple attributes derived from '{2}' defined on it. Only one such attribute is allowed.</target>
+        <note />
+      </trans-unit>
       <trans-unit id="UTA_NoValidConstructor">
         <source>Cannot find a valid constructor for test class '{0}'. Valid constructors are 'public' and either parameterless or with one parameter of type 'TestContext'.</source>
         <target state="translated">테스트 클래스 '{0}'에 대한 유효한 생성자를 찾을 수 없습니다. 유효한 생성자는 'public'이며 매개 변수가 없거나 'TestContext' 유형의 매개 변수가 하나 있습니다.</target>
@@ -360,11 +365,6 @@ Error: {1}</source>
 {2}</source>
         <target state="translated">테스트 메서드 {0}.{1}에 정의된 ExpectedException 특성이 생성하는 동안 예외를 throw했습니다.
 {2}</target>
-        <note></note>
-      </trans-unit>
-      <trans-unit id="UTA_MultipleExpectedExceptionsOnTestMethod">
-        <source>The test method {0}.{1} has multiple attributes derived from ExpectedExceptionBaseAttribute defined on it. Only one such attribute is allowed.</source>
-        <target state="translated">테스트 메서드 {0}.{1}에는 정의되어 있는 ExpectedExceptionBaseAttribute에서 파생된 특성이 여러 개 있습니다. 이러한 특성은 하나만 허용됩니다.</target>
         <note></note>
       </trans-unit>
       <trans-unit id="LegacyScenariosNotSupportedWarning">
